refactor: simplify shared cache configuration
