--- lms-setup/src/main/resources/application-qa.yaml
@@ -1,3 +1,0 @@
-setup:
-  security:
-    public-access: false
