--- lms-setup/src/main/resources/application-sandbox.yaml
@@ -1,3 +1,0 @@
-setup:
-  security:
-    public-access: false
