+++ shared-lib/shared-starters/starter-core/src/main/java/com/shared/starter_core/config/CoreExtrasAutoConfiguration.java
@@ -1,0 +1,9 @@
+package com.shared.starter_core.config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Import;
+
+@Configuration
+@Import({EnvironmentLogger.class, PerformanceConfig.class})
+public class CoreExtrasAutoConfiguration {
+}
