+++ shared-lib/shared-starters/starter-core/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,1 +1,2 @@
 com.common.starter.core.exception.CoreExceptionAutoConfiguration
+com.shared.starter_core.config.CoreExtrasAutoConfiguration
