+++ shared-lib/shared-starters/starter-redis/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,1 +1,2 @@
+com.shared.redis.starter.config.CacheAutoConfiguration
 com.shared.redis.starter.config.RedisAutoConfiguration
